+++ ja2lib/Strategic/MapScreenInterface.c
@@ -137,8 +137,8 @@ BOOLEAN fReBuildCharacterList = FALSE;
 INT32 giSizeOfInterfaceFastHelpTextList = 0;
 
 // Animated sector locator icon variables.
-INT16 gsSectorLocatorX;
-INT16 gsSectorLocatorY;
+u8 gsSectorLocatorX;
+u8 gsSectorLocatorY;
 UINT8 gubBlitSectorLocatorCode;    // color
 UINT32 guiSectorLocatorGraphicID;  // icon graphic ID
 // the animate time per frame in milliseconds
@@ -5084,14 +5084,14 @@ void TurnOnSectorLocator(UINT8 ubProfileID) {
       if (!fShowAircraftFlag) {
         // can't use his profile, he's where his chopper is
         Assert(iHelicopterVehicleId != -1);
-        gsSectorLocatorX = pVehicleList[iHelicopterVehicleId].sSectorX;
-        gsSectorLocatorY = pVehicleList[iHelicopterVehicleId].sSectorY;
+        gsSectorLocatorX = (u8)pVehicleList[iHelicopterVehicleId].sSectorX;
+        gsSectorLocatorY = (u8)pVehicleList[iHelicopterVehicleId].sSectorY;
       } else {
         return;
       }
     } else {
-      gsSectorLocatorX = gMercProfiles[ubProfileID].sSectorX;
-      gsSectorLocatorY = gMercProfiles[ubProfileID].sSectorY;
+      gsSectorLocatorX = (u8)gMercProfiles[ubProfileID].sSectorX;
+      gsSectorLocatorY = (u8)gMercProfiles[ubProfileID].sSectorY;
     }
   }
   gubBlitSectorLocatorCode = LOCATOR_COLOR_YELLOW;
